refactor: move middleware out
